requirements.txt
pytest==4.0.2
PyYAML==5.1
requests==2.20.1
